refactor: auth api, hooks
