+++ src/services/auth/apis/types.ts
@@ -1,0 +1,4 @@
+export interface Tokens {
+  accessToken: string;
+  refreshToken: string;
+}
