+++ src/services/auth/hooks/useSignIn.ts
@@ -1,0 +1,9 @@
+import { useMutation } from '@tanstack/react-query';
+
+import { signIn } from '~/services/auth/apis';
+
+export const useSignIn = () => {
+  return useMutation({
+    mutationFn: signIn,
+  });
+};
